minor cleanup and integrated alert res
