+++ app/src/main/res/values/styles.xml
@@ -12,5 +12,9 @@
         <item name="android:textSize"> 24sp </item>
         <item name="android:textColor"> #0E0E0E </item>
     </style>
-
+    <style name="MyAlertDialogTheme" parent="Theme.AppCompat.Dialog.Alert">
+        <item name="android:windowBackground">@drawable/popup_background</item>
+        <item name="android:textSize"> 17sp </item>
+        <item name="colorAccent">@color/colorAccent</item>
+    </style>
 </resources>
